// sourcecode/BTrak.Packages/snova-dashboard-module/Dashboard-module-components/dashboard-components/src/lib/snovasys-dashboard-management/models/hrDashboardModel.ts
export class HrDashboardModel {
    PageSize: number;
    PageNumber: number;
    SearchText: string;
    sortBy: string;
    sortDirectionAsc: boolean;
    selectedDate: string;
    type: string;
    branchId: string;
    workingStatus: string;
    teamLeadId: string;
    Date: string;
    isMorningLateEmployee: boolean;
    IsLunchBreakLongTake: boolean;
    IsMoreSpentTime: boolean;
    IsMorningAndAfterNoonLate: boolean;
    Order: string;
    entityId: string;
    DateFrom:string;
    DateTo:string;
    state: any;
 
}